Updates (#1)
* add version flag to cli

* update copyright to include 2020

* progress on standardizing user input

* remove argument from make_scenario

* Change gemfile source to http instead of https

* change download links from https to http

* Revert "change download links from https to http"

This reverts commit dc5563d27f662f9d6aea3b4398003fd54cc41e0a.

* update readme

* Add Gemfile.lock to gitignore

* update reminder comment to be more helpful

* remove Gemfile.lock from remote

* slight improvements to help text

* copy paste to get what is already in develop

* set version to 0.1.0

* Bugfix in delete

* Added SSL verification bypass

* call version better, and clean up copypasta

* Better readme for users and devs

* improve error help text

* Minor comment clarification

* Improve help text when making scenario files

* copy files from example project to cli repo

* Add Gemfile to be downloaded for users project folder

* Update create folder function to use files in cli

+++ lib/uo_cli.rb
@@ -1,7 +1,7 @@
 #!/usr/bin/ ruby
 
 #*********************************************************************************
-# URBANopt, Copyright (c) 2019, Alliance for Sustainable Energy, LLC, and other
+# URBANopt, Copyright (c) 2019-2020, Alliance for Sustainable Energy, LLC, and other
 # contributors. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without modification,
@@ -36,6 +36,7 @@ require "urbanopt/geojson"
 require "urbanopt/scenario"
 require "csv"
 require "json"
+require "openssl"
 
 
 module URBANopt
@@ -44,34 +45,35 @@ module URBANopt
     # Set up user interface
     @user_input = {}
     the_parser = OptionParser.new do |opts|
-        opts.banner = "Usage: uo [-pmradsf]\n" +
+        opts.banner = "Usage: uo [-pmradsfv]\n" +
         "\n" +
         "URBANopt CLI. \n" +
-        "For new projects, first create a project folder with -p, then run additional commands as desired \n" +
-        "For existing projects, specify your feature and scenarioCSV files to run (-r) and aggregate (-a) results"
-        
+        "First create a project folder with -p, then run additional commands as desired \n"
         opts.separator ""
 
-        opts.on("-p", "--project_folder <DIR>", "Create project directory named <DIR> in your current folder", String) do |folder|
+        opts.on("-p", "--project_folder <DIR>",String, "Create project directory named <DIR> in your current folder") do |folder|
             @user_input[:project_folder] = folder
         end
-        opts.on("-m", "--make_scenario <FFP>", "Create baseline ScenarioCSV file from <FFP> (Feature file path)", String) do |feature|
-            @user_input[:make_scenario_from] = feature
+        opts.on("-m", "--make_scenario", String, "Create ScenarioCSV files for each MapperFile using the Feature file path. Must specify -f argument") do
+            @user_input[:make_scenario_from] = "Create scenario files from FeatureFiles according to the MapperFiles in the 'mappers' directory"  # This text does not get displayed to the user
         end
-        opts.on("-r", "--run", "Run simulations. Must specify -s & -f arguments", String) do |run|
-            @user_input[:run_scenario] = "Run simulations"
+        opts.on("-r", "--run", String, "Run simulations. Must specify -s & -f arguments") do
+            @user_input[:run_scenario] = "Run simulations"  # This text does not get displayed to the user
         end
-        opts.on("-a", "--aggregate","Aggregate individual feature results to scenario-level results. Must specify -s & -f arguments", String) do |agg|
-            @user_input[:aggregate] = "Aggregate all features to a whole Scenario"
+        opts.on("-a", "--aggregate", String, "Aggregate individual feature results to scenario-level results. Must specify -s & -f arguments") do
+            @user_input[:aggregate] = "Aggregate all features to a whole Scenario"  # This text does not get displayed to the user
         end
-        opts.on("-d", "--delete_scenario <SFP>", "Delete results from scenario specified by <SFP> (ScenarioCSV file path)", String) do |delete|
-            @user_input[:delete_scenario] = delete
+        opts.on("-d", "--delete_scenario", String, "Delete results from scenario. Must specify -s argument") do
+            @user_input[:delete_scenario] = "Delete scenario results that were created from <SFP>"  # This text does not get displayed to the user
         end
-        opts.on("-s", "--scenario_file <SFP>", "Specify <SFP> (ScenarioCSV file path). Used when running and aggregating simulations", String) do |scenario|
+        opts.on("-s", "--scenario_file <SFP>", String, "Specify <SFP> (ScenarioCSV file path). Used as input for other commands") do |scenario|
             @user_input[:scenario] = scenario
         end
-        opts.on("-f", "--feature_file <FFP>", "Specify <FFP> (Feature file path). Used when running and aggregating simulations", String) do |feature|
+        opts.on("-f", "--feature_file <FFP>", String, "Specify <FFP> (Feature file path). Used as input for other commands") do |feature|
             @user_input[:feature] = feature
+        end
+        opts.on("-v", "--version", "Show CLI version and exit") do
+            @user_input[:version_request] = VERSION
         end
     end
 
@@ -80,9 +82,6 @@ module URBANopt
     rescue OptionParser::InvalidOption => e
       puts e
     end
-
-    # TODO: In newer versions of Ruby we can eliminate the need for each "do" block above by using this syntax. Have to see how run & agg work in that case
-    # end.parse!(into: @user_input)
 
 
     # Simulate energy usage for each Feature in the Scenario\
@@ -105,7 +104,6 @@ module URBANopt
         scenario_output = URBANopt::Scenario::ScenarioCSV.new(name, root_dir, run_dir, feature_file, mapper_files_dir, csv_file, num_header_rows)
         return scenario_output
     end
-
 
     # Create a scenario csv file from a FeatureFile
     # params\
@@ -134,7 +132,7 @@ module URBANopt
     # Includes weather for UO's example location, a base workflow file, and mapper files to show a baseline and a high-efficiency option.
     def self.create_project_folder(dir_name)
         if Dir.exist?(dir_name)
-             puts "ERROR:  there is already a directory here named #{dir_name}... aborting"
+            abort("ERROR:  there is already a directory here named #{dir_name}... aborting")
         else
             puts "CREATING URBANopt project directory: #{dir_name}"
             Dir.mkdir dir_name
@@ -143,32 +141,37 @@ module URBANopt
             mappers_dir_abs_path = File.absolute_path(File.join(dir_name, 'mappers/'))
             weather_dir_abs_path = File.absolute_path(File.join(dir_name, 'weather/'))
 
-            example_feature_file = "https://raw.githubusercontent.com/urbanopt/urbanopt-example-geojson-project/develop/example_project.json"
-            example_gem_file = "https://raw.githubusercontent.com/urbanopt/urbanopt-example-geojson-project/develop/Gemfile"
-            remote_mapper_files = ["https://raw.githubusercontent.com/urbanopt/urbanopt-example-geojson-project/develop/mappers/base_workflow.osw",
-                                   "https://raw.githubusercontent.com/urbanopt/urbanopt-example-geojson-project/develop/mappers/Baseline.rb",
-                                   "https://raw.githubusercontent.com/urbanopt/urbanopt-example-geojson-project/develop/mappers/HighEfficiency.rb"]
-            remote_weather_files = ["https://raw.githubusercontent.com/urbanopt/urbanopt-example-geojson-project/develop/weather/USA_NY_Buffalo-Greater.Buffalo.Intl.AP.725280_TMY3.epw",
-                                    "https://raw.githubusercontent.com/urbanopt/urbanopt-example-geojson-project/develop/weather/USA_NY_Buffalo-Greater.Buffalo.Intl.AP.725280_TMY3.ddy",
-                                    "https://raw.githubusercontent.com/urbanopt/urbanopt-example-geojson-project/develop/weather/USA_NY_Buffalo-Greater.Buffalo.Intl.AP.725280_TMY3.stat"]
+            # FIXME: When residential hpxml flow is implemented (https://github.com/urbanopt/urbanopt-example-geojson-project/pull/24 gets merged) these files will change
+            example_feature_file = "https://raw.githubusercontent.com/urbanopt/urbanopt-cli/updates/example_files/example_project.json"
+            example_gem_file = "https://raw.githubusercontent.com/urbanopt/urbanopt-cli/updates/example_files/Gemfile"
+            remote_mapper_files = [
+                "https://raw.githubusercontent.com/urbanopt/urbanopt-cli/updates/example_files/mappers/base_workflow.osw",
+                "https://raw.githubusercontent.com/urbanopt/urbanopt-cli/updates/example_files/mappers/Baseline.rb",
+                "https://raw.githubusercontent.com/urbanopt/urbanopt-cli/updates/example_files/mappers/HighEfficiency.rb",
+            ]
+            remote_weather_files = [
+                "https://raw.githubusercontent.com/urbanopt/urbanopt-cli/updates/example_files/weather/USA_NY_Buffalo-Greater.Buffalo.Intl.AP.725280_TMY3.epw",
+                "https://raw.githubusercontent.com/urbanopt/urbanopt-cli/updates/example_files/weather/USA_NY_Buffalo-Greater.Buffalo.Intl.AP.725280_TMY3.ddy",
+                "https://raw.githubusercontent.com/urbanopt/urbanopt-cli/updates/example_files/weather/USA_NY_Buffalo-Greater.Buffalo.Intl.AP.725280_TMY3.stat",
+            ]
             
             # Download files to user's local machine
             remote_mapper_files.each do |mapper_file|
                 mapper_root, mapper_base = File.split(mapper_file)
-                mapper_download = open(mapper_file)
+                mapper_download = open(mapper_file, {ssl_verify_mode: OpenSSL::SSL::VERIFY_NONE})
                 IO.copy_stream(mapper_download, File.join(mappers_dir_abs_path, mapper_base))
             end
             remote_weather_files.each do |weather_file|
                 weather_root, weather_base = File.split(weather_file)
-                weather_download = open(weather_file)
+                weather_download = open(weather_file, {ssl_verify_mode: OpenSSL::SSL::VERIFY_NONE})
                 IO.copy_stream(weather_download, File.join(weather_dir_abs_path, weather_base))
             end
-            gem_root, gem_base = File.split(example_gem_file)
-            example_gem_download = open(example_gem_file)
-            IO.copy_stream(example_gem_download, File.join(dir_name, gem_base))
+            # gem_root, gem_base = File.split(example_gem_file)
+            # example_gem_download = open(example_gem_file, {ssl_verify_mode: OpenSSL::SSL::VERIFY_NONE})
+            # IO.copy_stream(example_gem_download, File.join(dir_name, gem_base))
 
             feature_root, feature_base = File.split(example_feature_file)
-            example_feature_download = open(example_feature_file)
+            example_feature_download = open(example_feature_file, {ssl_verify_mode: OpenSSL::SSL::VERIFY_NONE})
             IO.copy_stream(example_feature_download, File.join(dir_name, feature_base))
         end
     end
@@ -179,57 +182,67 @@ module URBANopt
         create_project_folder(@user_input[:project_folder])
         puts "\nAn example FeatureFile is included: 'example_project.json'. You may place your own FeatureFile alongside the example."
         puts "Weather data is provided for the example FeatureFile. Additional weather data files may be downloaded from energyplus.net/weather for free"
-        puts "If you use additional weather files, ensure they are added to the 'weather' directory. You will need to configure your mapper file or your osw file to use the desired weather file"
-        puts "Next, move inside your new folder and create a baseline ScenarioFile using this CLI: 'uo -m'"
+        puts "If you use additional weather files, ensure they are added to the 'weather' directory. You will need to configure your mapper file and your osw file to use the desired weather file"
+        puts "Next, move inside your new folder ('cd <FolderYouJustCreated>') and create ScenarioFiles using this CLI: 'uo -m -f <FFP>'"
     end
 
     if @user_input[:make_scenario_from]
-        @feature_root, @feature_name = File.split(@user_input[:make_scenario_from])
-        puts "\nBuilding sample efficiency ScenarioFiles from #{@feature_name}..."
-        create_scenario_csv_file(@user_input[:make_scenario_from])
+        if @user_input[:feature].nil?
+            abort("\nYou must provide the '-s' flag and a valid path to a FeatureFile!\n---\n\n")
+        end
+        @feature_root, @feature_name = File.split(@user_input[:feature])
+        puts "\nBuilding sample ScenarioFiles, assigning mapper classes to each feature from #{@feature_name}..."
+        create_scenario_csv_file(@user_input[:feature])
+        puts "Done"
     end
 
     if @user_input[:run_scenario]
         if @user_input[:scenario].nil?
-            puts "\nYou must provide a valid path to a ScenarioFile!\n---\n"
-            raise OptionParser::MissingArgument 
+            abort("\nYou must provide '-s' flag and a valid path to a ScenarioFile!\n---\n\n")
         end
         if @user_input[:feature].nil?
-            puts "\nYou must provide a valid path to a FeatureFile!\n---\n"
-            raise OptionParser::MissingArgument
+            abort("\nYou must provide '-f' flag and a valid path to a FeatureFile!\n---\n\n")
         end
         @scenario_root, @scenario_name = File.split(@user_input[:scenario])
         @feature_root, @feature_name = File.split(@user_input[:feature])
-        puts "\nSimulating features of '#{@feature_name}' according to '#{@scenario_name}'..."
+        puts "\nSimulating features of '#{@feature_name}' as directed by '#{@scenario_name}'..."
         scenario_runner = URBANopt::Scenario::ScenarioRunnerOSW.new
         scenario_runner.run(run_func())
+        puts "Done"
     end
 
     if @user_input[:aggregate]
         if @user_input[:scenario].nil?
-            puts "\nYou must provide a valid path to a ScenarioFile!\n---\n"
-            raise OptionParser::MissingArgument 
+            abort("\nYou must provide '-s' flag and a valid path to a ScenarioFile!\n---\n\n")
         end
         if @user_input[:feature].nil?
-            puts "\nYou must provide a valid path to a FeatureFile!\n---\n"
-            raise OptionParser::MissingArgument
+            abort("\nYou must provide '-f' flag and a valid path to a FeatureFile!\n---\n\n")
         end
         @scenario_root, @scenario_name = File.split(@user_input[:scenario])
         @feature_root, @feature_name = File.split(@user_input[:feature])
         puts "\nAggregating results across all features of #{@feature_name} according to '#{@scenario_name}'..."
         scenario_result = URBANopt::Scenario::ScenarioDefaultPostProcessor.new(run_func()).run
         scenario_result.save
+        puts "Done"
     end
 
     if @user_input[:delete_scenario]
-        @scenario_root, @scenario_name = File.split(@user_input[:delete_scenario])
+        if @user_input[:scenario].nil?
+            abort("\nYou must provide '-s' flag and a valid path to a ScenarioFile!\n---\n\n")
+        end
+        @scenario_root, @scenario_name = File.split(@user_input[:scenario])
         scenario_name = @scenario_name.split('.')[0]
-        scenario_results_dir = File.join(@scenario_root, 'run', scenario_name)
+        scenario_root = File.absolute_path(@scenario_root)
+        scenario_results_dir = File.join(scenario_root, 'run', scenario_name)
         puts "\nDeleting previous results from '#{@scenario_name}'..."
         FileUtils.rm_rf(scenario_results_dir)
         puts "Done"
     end
 
-  end  # End CLI
+    if @user_input[:version_request]
+        puts "URBANopt CLI version: #{@user_input[:version_request]}"
+    end
 
-end  # End URBANopt
+  end  # End module CLI
+
+end  # End module Urbanopt
